Do not read stdout/stderr as binary data
When reading a file in binary mode, Ruby:
  1. Does not convert CR+LF to LF on windows;
  2. Force the encoding to ASCII-8BIT.

When attempting to print a String, ruby convert it to the encoding of
the teminal, generaly UTF-8, but arbitrary binary data cannot be
converted, and any non-ASCII char (with the most significant bit set to
1) will raise an Encoding::UndefinedConversionError exception, e.g.:

> "\xC3" from ASCII-8BIT to UTF-8 (Encoding::UndefinedConversionError)

Update the way we read stdout/stderr to avoid forcing encoding (we
assume the command was using the same encoding as the one we are reading
it's output with, and therefore that the output will be processable).
This fix the behavior of the shell agent when output contains
non-ASCCI-7BIT chars.